zirave-main/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
requests==2.31.0
aiofiles==23.2.1
google-generativeai==0.3.2